整合mac和linux的开发环境的配置

+++ git/install.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 
-
-
-ln -s $(pwd)/gitconfig ~/.gitconfig
-ln -s $(pwd)/gitignore_global ~/.gitignore_global
+if [[ ! -f ~/.gitconfig ]]; then
+  echo "configure git config"
+  ln -s $(pwd)/gitconfig ~/.gitconfig
+  ln -s $(pwd)/gitignore_global ~/.gitignore_global
+fi
